Single Channel Demo with current stimulation
Can now receive data over UDP and create a sine wave or square wave with arbitrary amplitude from received data.

Matlab demo is now made using UI elements to make it possible to send commands to the device using a dropdown menu to select the wave and a slider to select the amplitude.